refactor: move sign-in to view
