+++ src/app/models/loginResponse.ts
@@ -1,0 +1,3 @@
+export interface LoginResponse {
+  access_token: string;
+}
